chore(aya-obj): remove duplicate export of program items in aya-obj
The programs module in aya-obj has duplicate exported items that are
caused by using `pub mod *` in combination with `pub use *`. This
provides multiple ways to access the items.
For example `XdpAttachType`, may be accessed through:
- `aya_obj::programs::xdp::XdpAttachType`
- `aya_obj::programs::XdpAttachType`

This also causes duplicate entries in the public API file since it
records both public exported `mod`, and the public re-exported `use`.

To avoid this confusion and trim down the public api file while keeping
the same api, only the items of `aya_obj::programs::*` are publicly
exported, and not the sub-modules of program.