chore(aya-obj): remove duplicate export of obj items
The obj items are exported twice through `pub mod obj` and
`pub use obj::*`. This provides 2 ways to access the items.
For example:
- `aya_obj::copy_instructions`
- `aya_obj::obj::copy_instructions`

This also causes duplicate entries in the public api file due to the
items being exported twice.

To avoid this confusion and trim down the public api file while keeping
the same api, only the items of `obj` will be publicly exported, and not
the module.