chore(aya): remove obj alias to aya_obj
When `aya_obj` was migrated to be its own separate crate, the `obj`
alias to `aya_obj` was used so that the existing code would not break.
The use of the alias is a quick bandaid solution, however, it causes
some confusion when using `aya_obj` items since there are now 3 ways to
do this:
- `use aya_obj::*`
- `use obj::*`
- `use crate::obj::`

To avoid this confusion, the `obj` alias is removed so that items of
`aya_obj` are directly accessed through the full crate name.